moved game logic into cellGrid
